refactor: router 관련 상수 이동
